🐛 fix: 콜백 서스펜스 적용과 온보딩 경로 정리
